+++ src/main/java/net/brdle/collectorsreap/common/item/food/Nutrition.java
@@ -356,6 +356,13 @@ public class Nutrition {
 		.effect(() -> new MobEffectInstance(MobEffects.MOVEMENT_SPEED, 2400, 2), 1F)
 		.effect(() -> new MobEffectInstance(MobEffects.DIG_SPEED, 2400, 2), 1F)
 		.build();
+	public static final FoodProperties HEAVENS_CREAM = (new FoodProperties.Builder())
+		.alwaysEat()
+		.effect(() -> new MobEffectInstance(ModCompat.getTipsy(), 3600, 2), 1F)
+		.effect(() -> new MobEffectInstance(ModCompat.getIntoxication(), 2400, 0), 1F)
+		.effect(() -> new MobEffectInstance(ModCompat.getSweetHeart(), 2400, 2), 1F)
+		.effect(() -> new MobEffectInstance(CREffects.REBOUND.get(), 2400, 4), 1F)
+		.build();
 
 	public static final FoodProperties TIGER_PRAWN = (new FoodProperties.Builder())
 		.nutrition(2).saturationMod(0.1F)