+++ changelog.md
@@ -17,6 +17,7 @@
 - Added Lucuma Cake
 - Added Lucuma Ice Cream, Block, and Milkshake (for [Neapolitan](https://www.curseforge.com/minecraft/mc-mods/neapolitan))
 - Added Sunny Ice Cream (for [Neapolitan](https://www.curseforge.com/minecraft/mc-mods/neapolitan))
+- Added Heaven's Cream (for [Brewin' and Chewin'](https://www.curseforge.com/minecraft/mc-mods/brewin-and-chewin))
 - Added Lucuma Boat with Furnace and Large Lucuma Boat (for [Boatload](https://www.curseforge.com/minecraft/mc-mods/boatload))
 - Lunar Pearls can now be found in shipwreck and buried treasure chests
 - Strawberry Jam Bun now provides 8 nutrition and Instant Health